backend-app-api: split out ObservableConfigProxy
